[refactor]MainActivity
create FindAddress.class
->findEatingPlaceAddress() method
